// lib/types.ts
export interface GiftCard {
  id: string;
  name: string;
  brand: string;
  logo: string;
  availableValues: number[];
}

export interface User {
  name: string;
  nif: string;
  email: string;
  balance: number;
  initialBalance: number;
}

export interface Transaction {
  id: string;
  userId: string;
  cardId: string;
  amount: number;
  date: Date;
  status: 'pending' | 'confirmed' | 'refunded';
}

export interface CartItem {
  cardId: string;
  brand: string;
  value: number;
  quantity: number;
}


export interface UserCredentials {

  nif: string;

  email: string;

}



export interface UserValidationResponse {

  success: boolean;

  user?: {

    nif: string;

    email: string;

  };

  error?: string;

}